net: devmem: Split out net_devmem_binding_alloc and net_devmem_binding_populate
Instead of net_devmem_bind_dmabuf doing everything, separate it
into two steps:
- net_devmem_binding_alloc - allocate and map dmabuf
- net_devmem_binding_populate - populate genpool

Both are gonna be reused on the transmit side. No functional changes.

-----BEGIN COVER LETTER-----
net: devmem: TX Path

As discussed in [1], sending the preliminary version of the transmit
path. It is somewhat based on [2], but at this point probably nothing
of the original patch remains. Since we are not using fake pages
anymore, I had to do the following:
1. Add new bind_tx netlink to attach dmabufs to the netdev and export id
2. Ask users to pass that id in SCM_DEVMEM_DMABUF sendmsg
3. The payload chunks are referenced via zero-based iov_base and iov_len

1: https://lore.kernel.org/netdev/CAHS8izM8e4OhOFjRm9cF2LuN=ePWPgd-EY09fZHSybgcOaH4MA@mail.gmail.com/
2: https://lore.kernel.org/netdev/20230710223304.1174642-8-almasrymina@google.com/

v2:
- don't use iov_iter for tx chunks, it's slow
- drop net_devmem_dmabuf_lock
- remove loopback mode
- remove refcnt on tx (zc skb does sock_hold, can't detach binding from
  sock)
- separate bind-tx from bind-rx
- rename skb_add_rx_frag_netmem to skb_add_frag_netmem

Cc: Mina Almasry <almasrymina@google.com>
-----END COVER LETTER-----

Signed-off-by: Stanislav Fomichev <sdf@fomichev.me>