selftests: ncdevmem: Exercise chunked sendmsg
Currently the test is doing sendmsg with a single large 1M chunk. This
is the ideal and preferred case from the UAPI POW, but it doesn't exercise
a case where skb is constructed from smaller sub-PAGE_SIZE chunks.
Add an argument to specify max size per sendmsg and another test case
to use it.

Signed-off-by: Stanislav Fomichev <sdf@fomichev.me>