eth: bnxt: Remove most dependencies on RTNL
Only devlink and sriov paths are grabbing rtnl explicitly. The rest is covered
by netdev instance lock which the core now grabs, so there is no need to manage
rtnl explicitly anymore.

On the core side we can now try to drop rtnl in some places
(do_setlink for example) for the drivers that signal non-rtnl
mode (TBD).

Cc: Saeed Mahameed <saeed@kernel.org>
Signed-off-by: Stanislav Fomichev <sdf@fomichev.me>